postcare-api/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
openai
python-multipart
paddlepaddle
paddleocr
openai-whisper
edge-tts
Pillow
numpy
